+++ src/jaydee/crawlers/git.py
@@ -1,5 +1,6 @@
 from .base import BaseCrawler
 from jaydee import utils
+from jaydee.options import GitCrawlerOptions
 
 from tempfile import TemporaryDirectory
 import logging
@@ -20,10 +21,15 @@ class GitCrawler(BaseCrawler):
     """
 
     def __init__(
-        self, ignore=(".git", ".toml", ".lock", ".png", ".jpg", ".gif", ".gitignore")
+        self,
+        ignore=(".git", ".toml", ".lock", ".png", ".jpg", ".gif", ".gitignore"),
+        options=GitCrawlerOptions(),
     ):
         super().__init__()
+
+        # File extensions and directories to ignore when extracting data.
         self._ignore = ignore
+        self._options = options
 
     def extract_from_url(self, url: str) -> dict | None:
         """Extracts information from a repository given a valid clonable URL."""
@@ -31,6 +37,12 @@ class GitCrawler(BaseCrawler):
 
         # The result of the scraping process.
         result = None
+
+        # In strict mode check for repository validity before scraping.
+        if self._options._strict:
+            if not self.check_if_valid_repository(url):
+                logger.warning("Invalid git repository, skipping..")
+                return result
 
         logger.info(f"Scraping repository from URL: {url}")
         repo_name = url.rstrip("/").split("/")[-1]
@@ -73,3 +85,13 @@ class GitCrawler(BaseCrawler):
 
         logger.info("Finished scraping Github repository.")
         return result
+
+    def check_if_valid_repository(self, url: str) -> bool:
+        """Checks if the given url is a valid git repository or not."""
+        assert utils.validate_url(url), "The provided URL must be a valid url."
+
+        # Run the list remote to check if the url is a git repository.
+        result = subprocess.run(["git", "ls-remote", url])
+
+        # If it returns 0, the url is a valid git repository.
+        return result.returncode == 0